android-conversation-backup: update to 1.3 and major refactoring and clean up
Updated version to 1.3.1

Due to Google Play requirement that we support a target sdk version of 26,
The new version can only support a minimum sdk version of 14.
This corresponds to Android 4.0 generally available in 2012.

Streamlined processing so that permissions checking and capture is
seamlessly integrated.

Now, nothing happens until the user clicks "Back up", and at that
time, we will check permissions, store them, and determine
what actions to be taken by the service.

Use Intents Extras to pass data between activities and Service,
so that we don't depend on prior initialized values like
appdir and results log file.

There is now better error handling, as we show a dialog appropriately
when the external media is unavailable, or we cannot write to it,
or the permission to write to it wasn't granted by the user.

We also fixed issues so app works seamlessly on newer devices also
- Used the compatibility layer of Android APIs
- Fixed File URI permissions
- Fixed Notification support

Overall fixed up the code.