Use menu to show common functionality and refactor how dialogs are shown
Previously, home, archives and about were buttons on the main screen.
Since these are common, they have now been moved to the options menu.
This also reduces clutter on the activity screens, as the buttons there
are explicitly for the activity.

Previously, we showed dialogs using a deprecated showDialog() method on Activity.

Now, we use fragments and some shared code to unify how dialogs are created
and shown.